Add conesearch django app
