+++ daiquiri/conesearch/utils.py
@@ -1,0 +1,7 @@
+from django.conf import settings
+
+from daiquiri.core.utils import import_class
+
+
+def ConeSearchAdapter():
+    return import_class(settings.CONESEARCH_ADAPTER)()
